+++ srcs/ft_print_unsigned.c
@@ -6,7 +6,7 @@
 /*   By: mekherbo <mekherbo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/13 01:08:16 by mekherbo          #+#    #+#             */
-/*   Updated: 2023/11/15 18:42:28 by mekherbo         ###   ########.fr       */
+/*   Updated: 2023/11/16 00:01:46 by mekherbo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 int	ft_print_unsigned(va_list arg)
 {
-	unsigned long	u_nb;
+	unsigned int	u_nb;
 
-	u_nb = va_arg(arg, unsigned long);
+	u_nb = va_arg(arg, unsigned int);
 	ft_putnbr_base_unsigned((unsigned long long)u_nb, "0123456789");
-	return (ft_nblen_u(u_nb, 10));
+	return (ft_nblen_u((unsigned long)u_nb, 10));
 }
